requirements.txt
# Core Machine Learning
torch>=2.0
transformers>=4.38
datasets>=2.14
evaluate>=0.4.0
accelerate>=0.26.0

# Data handling
scikit-learn>=1.3
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
tqdm>=4.66

# (Optional but Recommended for Monitoring)
tensorboard>=2.14
wandb>=0.16

# (Optional if you want graphs later)
matplotlib>=3.8
